refactor: create topo-imagery-gdal package
