+++ packages/topo-imagery-gdal/src/topo_imagery_gdal/gdal/gdal_footprint.py
@@ -3,8 +3,8 @@ from decimal import Decimal
 
 from topo_imagery_common.files.files_helper import get_file_name_from_path
 
-from scripts.gdal.gdal_commands import get_footprint_command
-from scripts.gdal.gdal_helper import run_gdal
+from topo_imagery_gdal.gdal.gdal_commands import get_footprint_command
+from topo_imagery_gdal.gdal.gdal_helper import run_gdal
 
 SUFFIX_FOOTPRINT = "_footprint.geojson"
