+++ packages/topo-imagery-gdal/src/topo_imagery_gdal/gdal/gdal_presets.py
@@ -5,7 +5,7 @@ from typing import Annotated
 DEFAULT_NO_DATA_VALUE: Annotated[Decimal, "From the New Zealand National Aerial LiDAR Base Specification"] = Decimal(-9999)
 
 SCALE_254_ADD_NO_DATA = ["-scale", "0", "255", "0", "254", "-a_nodata", "255"]
-""" Scale imagery from 0-255 to 0-254 then set 255 as NO_DATA. 
+""" Scale imagery from 0-255 to 0-254 then set 255 as NO_DATA.
 Useful for imagery that does not have a alpha band.
 """
 BASE_COG = [
