+++ packages/topo-imagery-gdal/test/gdal/gdalinfo_test.py
@@ -1,6 +1,6 @@
-from scripts.gdal.gdalinfo import get_origin
-from scripts.gdal.tests.gdalinfo import fake_gdal_info
-from scripts.tile.tile_index import Point
+from topo_imagery_gdal.gdal.gdalinfo import get_origin
+from topo_imagery_gdal.testing.gdalinfo import fake_gdal_info
+from topo_imagery_gdal.tile.tile_index import Point
 
 
 def test_get_origin() -> None:
