+++ scripts/generate_hillshade.py
@@ -12,10 +12,10 @@ from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT
 from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType, is_tiff
 from topo_imagery_common.files.fs import exists, read, write, write_all
 from topo_imagery_common.log.time_helper import time_in_ms
+from topo_imagery_gdal.gdal.gdal_commands import get_gdal_command, get_hillshade_command
+from topo_imagery_gdal.gdal.gdal_helper import run_gdal
+from topo_imagery_gdal.gdal.gdal_presets import CompressionPreset, HillshadePreset
 
-from scripts.gdal.gdal_commands import get_gdal_command, get_hillshade_command
-from scripts.gdal.gdal_helper import run_gdal
-from scripts.gdal.gdal_presets import CompressionPreset, HillshadePreset
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.create_stac import create_item
 from scripts.standardising import create_vrt
