+++ scripts/stac/imagery/capture_area.py
@@ -9,25 +9,7 @@ from shapely.geometry import MultiPolygon
 from shapely.geometry.base import BaseGeometry
 from shapely.ops import orient
 from topo_imagery_common.log.time_helper import time_in_ms
-
-DECIMAL_DEGREES_1M = Decimal("0.00001")
-"""
-Degree precision of ~1m (decimal places 5, https://en.wikipedia.org/wiki/Decimal_degrees)
-"""
-
-
-def get_buffer_distance(gsd: Decimal) -> float:
-    """The `gsd` (in meters) is multiplied by 2 and then by the 1m degree of precision.
-    A `buffer factor` of 2 was decided on after experimenting with different outputs,
-    details of this can be found in TDE-1049.
-
-    Args:
-        gsd: Ground Sample Distance in meters
-
-    Returns:
-        buffer distance as a float
-    """
-    return float(gsd * 2 * DECIMAL_DEGREES_1M)
+from topo_imagery_gdal.gdal.gdal_commands import get_buffer_distance
 
 
 def to_feature(geometry: BaseGeometry) -> dict[str, Any]:
