+++ scripts/stac/imagery/tests/create_stac_test.py
@@ -5,9 +5,9 @@ from typing import Any, cast
 
 from pytest_subtests import SubTests
 from topo_imagery_common.datetimes import format_rfc_3339_datetime_string
+from topo_imagery_gdal.gdal.gdalinfo import GdalInfo
 
 from scripts.conftest import any_epoch_datetime, any_epoch_datetime_string
-from scripts.gdal.gdalinfo import GdalInfo
 from scripts.stac.imagery.collection import ImageryCollection
 from scripts.stac.imagery.collection_context import CollectionContext
 from scripts.stac.imagery.create_stac import create_collection, create_item, get_items_to_replace, merge_item_list_for_resupply
